+++ src/ComponentCollection.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rougin\Slytherin;
+
+/**
+ * Component Collection
+ *
+ * Contains all the required components for Slytherin.
+ * 
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class ComponentCollection extends Components {}